+++ test/dummy/app/controllers/posts_controller.rb
@@ -1,0 +1,5 @@
+class PostsController < ApplicationController
+  def show
+    @post = present Post.new
+  end
+end