+++ test/dummy/app/presenters/post_presenter.rb
@@ -1,0 +1,5 @@
+class PostPresenter < Oprah::Presenter
+  def paragraph(&block)
+    h.content_tag(:p, &block)
+  end
+end