+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   argument, which takes either a Class or an Array of classes
 - Replace repeated method default arguments with splats
 - Add assertions to TestHelper
+- Delegate to most recent view context in presenters created
+  from within Rails controllers [#1]
 
 0.1.2
 -----